# api/docs/swagger.yaml
definitions:
...
  models.ListMedicineWarehouse:
    properties:
      count:
        type: integer
      medicine_warehouses:
        items:
          $ref: '#/definitions/models.MedicineWarehouse'
        type: array
    type: object
  models.MedicineWarehouse:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      quantity:
        type: integer
      quantity_type:
        type: string
      updated_at:
        type: string
    type: object
  models.MedicineWarehouseCreate:
    properties:
      name:
        type: string
      quantity:
        type: integer
      quantity_type:
        type: string
    type: object
  models.ResponseError:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  models.StatusRes:
    properties:
      status:
        type: boolean
    type: object
...
  models.UpdateMedicineWarehouseReq:
    properties:
      id:
        type: string
      name:
        type: string
      quantity:
        type: integer
      quantity_type:
        type: string
    type: object
host: localhost:9050
info:
  contact: {}
  title: CRM Farmish
  version: "1.7"
paths:
...
  /v1/medicine:
    delete:
      consumes:
      - application/json
      description: DeleteMedicineWarehouse - Api for Delete Medicine Warehouse
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StatusRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Delete Medicine Warehouse
      tags:
      - Medicine Warehouse
    get:
      consumes:
      - application/json
      description: ListMedicineWarehouse - Api for List Medicine Warehouse
      parameters:
      - in: query
        name: delete_status
        type: boolean
      - in: query
        name: limit
        type: integer
      - in: query
        name: order_by
        type: string
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListMedicineWarehouse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: List Medicine Warehouse
      tags:
      - Medicine Warehouse
    post:
      consumes:
      - application/json
      description: CreateMedicineWarehouse - Api for crete MedicineWarehouse
      parameters:
      - description: AnimalTypeCreate
        in: body
        name: AnimalTypeCreate
        required: true
        schema:
          $ref: '#/definitions/models.MedicineWarehouseCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MedicineWarehouse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create Medicine Warehouse
      tags:
      - Medicine Warehouse
    put:
      consumes:
      - application/json
      description: UpdateMedicineWarehouse - Api for Update Medicine Warehouse
      parameters:
      - description: UpdateAnimalTypeReq
        in: body
        name: UpdateAnimalTypeReq
        required: true
        schema:
          $ref: '#/definitions/models.UpdateMedicineWarehouseReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MedicineWarehouse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update Medicine Warehouse
      tags:
      - Medicine Warehouse
  /v1/medicine/get:
    get:
      consumes:
      - application/json
      description: GetMedicineWarehouse - Api for Get Medicine Warehouse
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MedicineWarehouse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get Medicine Warehouse
      tags:
      - Medicine Warehouse
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"